% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make.Tree.R
\name{Deep.Compute}
\alias{Deep.Compute}
\title{Compute the deep position of every node}
\usage{
Deep.Compute(tree)
}
\arguments{
\item{tree}{igraph object representing the tree}
}
\value{
igraph object updated
}
\description{
For the computation of the complete decision model is necessary to stablish the 
deep position of every node.
}
\author{
Pedro Guarderas, Andrés Lopez
}
\seealso{
\code{\link{Read.Tree}}
}

