% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tree.R
\name{deep_compute}
\alias{deep_compute}
\title{Compute the deep position of every node}
\usage{
deep_compute(tree)
}
\arguments{
\item{tree}{igraph object representing the tree}
}
\value{
igraph object updated
}
\description{
For the computation of the complete decision model is necessary to establish the 
deep position of every node.
}
\seealso{
\code{\link{Read.Tree}}
}
\author{
Pedro Guarderas, Andrés Lopez
}
