% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_criteria.R
\name{eval_criteria}
\alias{eval_criteria}
\title{Evaluation of decision tree nodes}
\usage{
eval_criteria(tree, weights, utilidades, rescale = FALSE, indexto = TRUE,
  idcolname = TRUE)
}
\arguments{
\item{tree}{inicial tree struture with utilities in its leafs}

\item{weights}{weights}

\item{utilities}{inicial tree struture with utilities in its leafs}

\item{types}{the different names for the levels in the tree, the most common names are: 'root',
'criteria', 'subcriteria', 'utility'}
}
\value{
data.table with with decision model computed
}
\description{
Evaluation of decision tree nodes
}
\details{
Details
}
\examples{
x<-2
}
\author{
Pedro Guarderas, Andrés Lopez
}
\seealso{
\code{\link{read_weights}}, \code{\link{eval_index}}
}

