% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Export.Tree.R
\name{Export.Decision.Tree}
\alias{Export.Decision.Tree}
\title{Compute decision model for tree}
\usage{
Export.Decision.Tree(tree, utilities, weights, types = c("root", "criteria",
  "subcriteria", "utility"))
}
\arguments{
\item{tree}{inicial tree struture with utilities in its leafs}

\item{utilities}{inicial tree struture with utilities in its leafs}

\item{weights}{weights}

\item{types}{the different names for the levels in the tree, the most common names are: ROOT, 
CRITERIA, SUBCRITERIA, UTILITY}
}
\value{
data.table with with decision model computed
}
\description{
The decision tree for the model base in MAUT has nodes with associated utilities, 
this function compute all the internal nodes utilities.
}
\details{
Details
}
\examples{
x<-2
}
\author{
Pedro Guarderas, Andrés Lopez
}
\seealso{
\code{\link{read_weights}}, \code{\link{eval_index}}
}

