% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make.Tree.R
\name{Index.Weights}
\alias{Index.Weights}
\title{Compute leaves weights}
\usage{
Index.Weights(tree)
}
\arguments{
\item{tree}{igraph object representing the tree}
}
\value{
igraph object updated
}
\description{
The computation of weights could be determined in an inverse processes given the
internal weights.
}
\seealso{
\code{\link{Read.Tree}}
}
\author{
Pedro Guarderas, Andrés Lopez
}
