% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tree.R
\name{index_weights}
\alias{index_weights}
\title{Compute leaves weights}
\usage{
index_weights(tree)
}
\arguments{
\item{tree}{igraph object representing the tree}
}
\value{
igraph object updated
}
\description{
The computation of weights could be determined in an inverse processes given the
internal weights.
}
\seealso{
\code{\link{read_tree}}
}
\author{
Pedro Guarderas, Andrés Lopez
}
