% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make.Tree.R
\name{Make.Decision.Tree}
\alias{Make.Decision.Tree}
\title{Evaluate utilities}
\usage{
Make.Decision.Tree(tree.data)
}
\arguments{
\item{tree.data}{tree}
}
\value{
igraph structure
}
\description{
Create decision tree for MAUT models exporting to igraph struture
}
\details{
Details
}
\examples{
x<-1
Make.Decision.Tree( tree.data )
}
\author{
Pedro Guarderas, Andrés Lopez
}
\seealso{
\code{\link{Read.Tree}}
}

