% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelo.R
\name{modelo.calcula.tree}
\alias{modelo.calcula.tree}
\title{Compute decision model for tree}
\usage{
modelo.calcula.tree(tree, tipo)
}
\arguments{
\item{tree}{inicial tree struture with utilities in its leafs}

\item{tipo}{the different names for the levels in the tree, the most common names are: ROOT, 
CRITERIA, SUBCRITERIA, UTILITY}
}
\value{
Returns data table with definition of utility functions by range
}
\description{
The decision tree for the model base in MAUT has nodes with associated utilities, 
this function compute all the internal nodes utilities.
}
\details{
Details
}
\examples{
x<-2
}
\author{
Pedro Guarderas, Andrés Lopez
}
\seealso{
\code{\link{read_weights}}, \code{\link{eval_index}}
}

