% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofensive_defensive.R
\name{od_rating}
\alias{od_rating}
\title{Offensive - Defensive rating method}
\usage{
od_rating(A, iter = 1000, rer = 1e-12)
}
\arguments{
\item{A}{matrix with scores}

\item{iter}{number of iterations}

\item{rer}{relative error to stop computation}
}
\value{
list with rating, offensive score, defensive score and number of iterations.
}
\description{
Computes rating using the offensive-defensive method
}
\examples{
A <- matrix( c( 0, 7, 21, 7, 0,
52, 0, 34, 25, 27, 
24, 16, 0, 7, 3,
38, 17, 5, 0, 14,
45, 7, 30, 52, 0 ), nrow = 5, ncol = 5 )
od_rating( A ) 
}
\author{
Pedro Guarderas
\email{pedro.felipe.guarderas@gmail.com}
}
