% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.Simulation.Weight.R
\name{Plot.Simulation.Weight}
\alias{Plot.Simulation.Weight}
\title{Simulation of constrained weights}
\usage{
Plot.Simulation.Weight(S, title = "Simulations", xlab = "ID",
  ylab = "Utility", lines.cols = "blue", box.col = "gold",
  box.outlier.col = "darkred", utility.col = "darkgreen",
  utility.point.col = "darkgreen", text.col = "black")
}
\arguments{
\item{S}{simulation}

\item{title}{title for plot}

\item{xlab}{label x-axis}

\item{ylab}{label y-axis}

\item{lines.cols}{spectrum of colors for simulations}

\item{box.col}{box color}

\item{box.outlier.col}{box outlier color}

\item{utility.col}{color for utility value}

\item{utility.point.col}{color for point utility value}

\item{text.col}{text color}
}
\value{
ggplot object of simulation
}
\description{
Simulation of weight following a Dirichlet distribution with a concentration 
parameter and satisfying sum constraints.
}
\details{
Details
}
\author{
Pedro Guarderas
}
\seealso{
\code{\link{Sim.Const.Weights}} \code{\link{Sim.Weights}}
}

