% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spider.Plot.R
\name{Spider.Plot}
\alias{Spider.Plot}
\title{Spider plot}
\usage{
Spider.Plot(data, data.label, data.fill, data.color, data.linetype, data.alpha,
  data.size, data.label.color, data.label.size, group, criteria, valor, title,
  title.color, title.size, label.size, label.color, label.angle, label.position,
  theta, grid, grid.color, grid.radius.color, grid.linetype, grid.size,
  grid.radius.linetype, grid.radius.size, axis, axis.label, axis.color,
  axis.size, axis.linetype, axis.angle, axis.label.color, axis.label.size,
  axis.label.displace, axis.label.angle, legend.position, legend.size,
  legend.text.color, plot.margin)
}
\arguments{
\item{data}{Datos con los valores a plotear, data.frame}

\item{data.label}{label}

\item{data.fill}{fill}

\item{data.color}{color}

\item{data.linetype}{linetype}

\item{data.alpha}{alpha}

\item{data.size}{size}

\item{data.label.color}{label.color}

\item{data.label.size}{label.size}

\item{group}{Columna de data por la cual se agrupan los datos}

\item{criteria}{Columna con los criterios de evaluación, debe ser}

\item{valor}{Columna con los valores de los polígonos}

\item{title}{Título para el plot}

\item{title.color}{Color del título}

\item{title.size}{Tamaño del título}

\item{label.size}{Tamaño de etiquetas}

\item{label.color}{Color de etiquetas}

\item{label.angle}{Ángulo de etiqueas}

\item{label.position}{Posición de etiqueas}

\item{theta}{Giro del plot}

\item{grid}{Malla para los radios, vector numérico con valores únicos}

\item{grid.color}{Color del grid}

\item{grid.radius.color}{Color de los radios del grid}

\item{grid.linetype}{Tipo de línea del grid}

\item{grid.size}{Tamaño de la línea}

\item{grid.radius.linetype}{Tipo de línea de los radios del grid}

\item{grid.radius.size}{Tamaño de línea de los radios}

\item{axis}{Eje para ubicarse en el grid}

\item{axis.label}{Etiquetas del eje}

\item{axis.color}{Color del eje}

\item{axis.size}{Tamaño del eje}

\item{axis.linetype}{Tipo del eje}

\item{axis.angle}{Ángulo del eje}

\item{axis.label.color}{Color para labels}

\item{axis.label.size}{Tamaño del texto del eje}

\item{axis.label.displace}{Desplazamiento de los labels con respecto al eje}

\item{axis.label.angle}{Ángulo de labels}

\item{legend.position}{position for the legend}

\item{legend.size}{legend size}

\item{legend.text.color}{legend text color}

\item{plot.margin}{Márgenes del plot}
}
\value{
data.table with utilities
}
\description{
Spider plot
}
\details{
Details
}
\examples{
# Preparing data
n<-10
m<-7
cols<-sample( colors()[ grepl('(red|blue|olive|darkgree)', colors() ) ], m, replace = TRUE )

data<-data.frame( grp = paste( 'A', sort( rep( 1:m, n ) ), sep = '' ),
                  cri = factor( rep( paste( 'c', 1:n, sep = '' ), m ), 
                                levels = paste( 'c', 1:n, sep = '' ), ordered = TRUE ),
                  val = runif( m * n ) )

data.label<-paste( 'A', 1:m,  ' class', sep = '' )
data.fill<-cols
data.color<-cols
data.linetype<-rep( 'solid', m )
data.alpha<-rep( 0.05, m )
data.size<-rep( 0.7, m )
data.label.color<-'black'
data.label.size<-15

# Spider plot parameters
title<-'Spider'
title.color<-'red3'
title.size<-20

label.size<-rep( 8, n )
label.color<-rep( 'steelblue4', n )
label.angle<-rep( 0, n )
label.position<-rep( 1.1, n )

theta<-pi/2

grid<-sort( c( 0.1, 0.25, 0.5, 0.75, 1.0 ) )
grid.color<-'grey'
grid.radius.color<-'dodgerblue3'
grid.linetype<-'dashed'
grid.size<-0.5
grid.radius.linetype<-'solid'
grid.radius.size<-0.5

axis<-grid # Same as grid
axis.label<-paste( 100 * axis, '\%', sep = '' )
axis.color<-'black'
axis.size<-0.7
axis.linetype<-'solid'
axis.angle<-0.4*pi
axis.label.color<-'darkgreen'
axis.label.size<-5
axis.label.displace<- -0.07
axis.label.angle<-0

legend.position<-c(0.9, 0.9)
legend.size<-0.5
legend.text.color<-'black'

plot.margin<-unit( c( 1.0, 1.0, 1.0, 1.0 ),"cm")


p<-Spider.Plot( data,
                data.label,
                data.fill,
                data.color,
                data.linetype,
                data.alpha,
                data.size,
                data.label.color, 
                data.label.size,
                grp,
                cri,
                val, 
                title,
                title.color,
                title.size,
                label.size, 
                label.color,
                label.angle,
                label.position,
                theta, 
                grid, 
                grid.color, 
                grid.radius.color,
                grid.linetype, 
                grid.size, 
                grid.radius.linetype, 
                grid.radius.size, 
                axis, 
                axis.label,
                axis.color,
                axis.size, 
                axis.linetype, 
                axis.angle, 
                axis.label.color,
                axis.label.size, 
                axis.label.displace,
                axis.label.angle,
                legend.position,
                legend.size,
                legend.text.color,
                plot.margin )

plot(p)
}
\author{
Pedro Guarderas, Andrés Lopez
}

