% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make.Tree.R
\name{Sum.Weights}
\alias{Sum.Weights}
\title{Sum weights for internal nodes}
\usage{
Sum.Weights(tree)
}
\arguments{
\item{tree}{igraph object representing the tree}
}
\value{
igraph object updated
}
\description{
The weights of the internal nodes has to be computed first is necessary to add each
weights of the leaves.
}
\seealso{
\code{\link{Read.Tree}}
}
\author{
Pedro Guarderas, Andrés Lopez
}
