% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weight.Sensitivity.R
\name{Weight.Sensitivity}
\alias{Weight.Sensitivity}
\title{Sensitivity to weight}
\usage{
Weight.Sensitivity(utilities, weights)
}
\arguments{
\item{utilities}{utilities}

\item{weights}{weights}
}
\value{
var
}
\description{
Create decision tree for MAUT models exporting to igraph struture
}
\details{
Details
}
\examples{
Weight.Sensitivity( utilidades, weights )
}
\author{
Pedro Guarderas, Andrés Lopez
}
\seealso{
\code{\link{Eval.Utilities}}
}

