% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rating_methods.R
\name{borda_count}
\alias{borda_count}
\title{Borda count}
\usage{
borda_count(R, v = NULL, u = NULL)
}
\arguments{
\item{R}{matrix with rankings}

\item{v}{vector of votes for each ranking}

\item{u}{vector of weight for each ranking}
}
\value{
Vector with aggregated ranking
}
\description{
Rank aggregation with the Borda count method
}
\examples{
R <- matrix( c( 'A', 'B', 'C', 'B', 'A', 'C', 'C', 'A', 'B' ), nrow = 3, ncol = 3 )
v <- c( 5, 4, 3 )
r <- borda_count( R, v )
}
\author{
Pedro Guarderas
\email{pedro.felipe.guarderas@gmail.com}
}
