% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rating_methods.R
\name{borda_count}
\alias{borda_count}
\title{Borda count}
\usage{
borda_count(R)
}
\arguments{
\item{R}{matrix with rankings}
}
\value{
Vector with aggregated ranking
}
\description{
Rank aggregation with the Borda count method
}
\author{
Pedro Guarderas
\email{pedro.felipe.guarderas@gmail.com}
}
