% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borda_count.R
\name{borda_count}
\alias{borda_count}
\title{Borda count}
\usage{
borda_count(R)
}
\arguments{
\item{R}{matrix with rakings}
}
\value{
Vector with aggregated raking
}
\description{
Rank aggreation with the Borda count method
}
\author{
Pedro Guarderas
\email{pedro.felipe.guarderas@gmail.com}
}
