% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colley_method.R
\name{colley_rating}
\alias{colley_rating}
\title{Colley method}
\usage{
colley_rating(n, g, w, l, t = NULL)
}
\arguments{
\item{n}{symmetric matrix of number of times each player faced another, zero diagonal}

\item{g}{total number of games played by each player}

\item{w}{accumulated vector of wins for each player}

\item{l}{accumulated vector of losses for each player}

\item{t}{symmetric matrix of ties, with zero diagonal}
}
\value{
Vector with ratings
}
\description{
Rank computation using the Cooley method with ties if required
}
\examples{
# Example without ties
n <- matrix( 1, 5, 5 )
diag( n ) <- 0
g <- rep( 4, 5 )
w <- c( 0, 4, 0, 0, 2 )
l <- c( 4, 0, 0, 2, 0 )
r <- colley_rating( n, g, w, l )
}
\author{
Pedro Guarderas
\email{pedro.felipe.guarderas@gmail.com}
}
