% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute.Model.R
\name{Compute.Model}
\alias{Compute.Model}
\title{Evaluation of decision tree nodes}
\usage{
Compute.Model(tree, utilities, weights)
}
\arguments{
\item{tree}{inicial tree struture with utilities in its leafs}

\item{utilities}{data.table with ordered columns containing the values of utilities}
}
\value{
data.table with with decision model computed
}
\description{
Evaluation of decision tree nodes
}
\details{
Details
}
\examples{
# Index
index<-data.table( cod = c( 'A', 'B', 'C', 'D' ), 
                   i1 = c( 0.3428570, 1, 1, 1 ),
                   i2 = c( 0.5, 0.5, 1, 0.5 ), 
                   i3 = c( 0.5, 1.0, 0.75, 0.25 ),
                   i4 = c( 0, 0.2696746, 0.6751261, 0.7401660 ),
                   i5 = c( 0.2797259, 0.2981198, 1, 0.1952864 ) )

# Loading utilities
file<-'example/utilities.txt'
script<-'utilities.R'
lines<-17
skip<-2
encoding<-'utf-8'
functions<-Read.Utilities( file, script, lines, skip, encoding )
source( 'utilities.R' )

# Index positions
columns<-c( 2, 3, 5, 6 )

# Associated names of functions
functions<-sapply( c( 'Project', 'Self implementation', 'External and local relations', 
                      'Scope of capabilities' ),
                   FUN = Stand.String )
names( functions )<-NULL

# Evaluation of utilities
utilities<-Eval.Utilities( index, columns, functions )

# Make tree
file<-'example/tree.xls'
sheetIndex<-1
cols<-1:5
rows<-c(1,8)
tree.data<-Read.Tree( file, sheetIndex, cols, rows )
tree<-Make.Decision.Tree( tree.data )

# Compute the decision model
weights<-tree.data[ !is.na( weight ) ]$weight
model<-Compute.Model( tree, utilities, weights )
}
\author{
Pedro Guarderas, Andrés Lopez
}
\seealso{
\code{\link{read_weights}}, \code{\link{eval_index}}
}

