% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tree.R
\name{divide_weights}
\alias{divide_weights}
\title{Divide weights of internal nodes}
\usage{
divide_weights(tree)
}
\arguments{
\item{tree}{igraph object representing the tree}
}
\value{
igraph object updated
}
\description{
After the addition of weights for internal nodes the final weights have to be 
computed dividing by the total weight of each parent.
}
\seealso{
\code{\link{read_tree}}
}
\author{
Pedro Guarderas, Andrés Lopez
}
