% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eval.Utilities.R
\name{Eval.Utilities}
\alias{Eval.Utilities}
\title{Evaluate utilities}
\usage{
Eval.Utilities(index, names, indexsub, colcod = 1, colpos, colfun, coltip,
  envir = .GlobalEnv)
}
\arguments{
\item{index}{data of indexes}

\item{names}{data with name of indexes}

\item{indexsub}{prefijo del nombre de las columnas de los indicadores.}

\item{colcod}{columna de inicio del data frame \code{index} definido previamente.}

\item{colpos}{columna de inicio del data frame \code{names} definido previamente.}

\item{colfun}{columna que indica el nombre interno de las funciones de utilidad.}

\item{coltip}{columna que indica el tipo de variable de los indicadores, cualitativos o 
cuantitativos.}

\item{envir}{entorno de trabajo global, definido con las funciones de utilidad.}
}
\value{
La función \code{eval_index} retorna un data frame que contiene los resultados de los 
indicadores al ser evaluados por las funciones de utilidad definidas.
}
\description{
Evaluaction of utilities for a data.frame of indexes
}
\details{
Details
}
\examples{
# Indicadores
indic<-data.frame(cod=c('A', 'B', 'C', 'D'), 
                  i1=c(0.3428570, 1, 1, 1),
                  i2=c(0.5, 0.5, 1, 0.5), 
                  i3=c(0, 0.2696746, 0.6751261, 0.7401660),
                  i4=c(0.2797259, 0.2981198, 1, 0.1952864))
colnames(indic)

# Nombre de indicadores
nomb_indic<-data.frame(cod=1:4, 
                    nom0=c('PROYECTO INSTITUCIONAL', 'SEGUIMIENTO IMPLEMENTADO',
                           'PROYECTOS DE VINCULACION', 'ACTIVIDADES DE EXTENSION Y DIVULGACION'),
                    nom1=c('Proyecto institucional', 'Seguimiento implementado', 
                           'Proyectos de vinculación', 'Actividades de extensión y divulgación'),
                    nomfun=c('proyecto_institucional', 'seguimiento_implementado', 
                             'proyectos_vinculacion', 'actividades_extension_divulgacion'),
                    tip=c('CUALITATIVA', 'CUALITATIVA', 'CUANTITATIVA', 'CUANTITATIVA'))

# Script R con las funciones de utilidad.
source('example/funciones_utilidad.R')

# Uso de la funcion eval_index
indic_eval<-Eval.Utilities( index = indic, names = nomb_indic, indexsub = 'i',
                     colcod = 1, colpos = 1, colfun = 4, coltip = 5 )
}
\author{
Julio Andrade, Pedro Guarderas, Daniel Lagos, Andrés López, Paúl Recalde, Edison Salazar
}
\seealso{
\code{\link{read_weights}}, \code{\link{eval_index}}
}

