% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make.Tree.R
\name{Make.Decision.Tree}
\alias{Make.Decision.Tree}
\title{Evaluate utilities}
\usage{
Make.Decision.Tree(tree.data)
}
\arguments{
\item{tree.data}{tree}
}
\value{
igraph structure
}
\description{
Create decision tree for MAUT models exporting to igraph struture
}
\details{
Details
}
\examples{
file<-'example/tree.xls'
sheetIndex<-1
cols<-1:5
rows<-c(1,8)
tree.data<-Read.Tree( file, sheetIndex, cols, rows )
tree<-Make.Decision.Tree( tree.data )
plot(tree)
}
\seealso{
\code{\link{Read.Tree}}
}
\author{
Pedro Guarderas, Andrés Lopez
}
