\name{mau-package}
\alias{mau-package}
\alias{mau}
\docType{package}

\title{
Rutinas para Modelos de Decision y Evaluacion.
}
\description{
Paquete que contiene un conjunto de rutinas para generar modelos decision.
}
\details{
\tabular{ll}{
Package: \tab mau\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-12-17\cr
License: \tab LGPL-3\cr
}
}
\author{
Felipe Aguirre, Julio Andrade, Daniel Lagos, Andrés López, Pedro Guarderas, Paúl Recalde, 
Edison Salazar.
Maintainer: <pedro.felipe.guarderas@gmail.com>
}
\references{
}

\keyword{ package }
\seealso{
\code{}
}
\examples{
}
