% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.Tree.R
\name{Read.Tree}
\alias{Read.Tree}
\title{Evaluate utilities}
\usage{
Read.Tree(file, skip, nrows)
}
\arguments{
\item{file}{file}

\item{skip}{index}

\item{nrows}{rows}
}
\value{
data.table with utilities
}
\description{
Read decision tree for MAUT model
}
\details{
Details
}
\examples{
file<-'example/tree.csv'
sheetIndex<-1
cols<-1:5
rows<-c(1,8)
tree.data<-Read.Tree( file, sheetIndex, cols, rows )
}
\seealso{
\code{\link{Read.Utilities}}, \code{\link{Make.Decision.Tree}}
}
\author{
Pedro Guarderas, Andrés Lopez
}
