% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.Tree.R
\name{Read.Tree}
\alias{Read.Tree}
\title{Evaluate utilities}
\usage{
Read.Tree(file, sheetIndex, cols, rows)
}
\arguments{
\item{file}{file}

\item{cols}{cols}

\item{rows}{rows}

\item{index_file}{index}

\item{sheet}{sheet}

\item{cols_index}{cols_index}
}
\value{
data.table with utilities
}
\description{
Read decision tree for MAUT model
}
\details{
Details
}
\examples{
Read.Tree( file, index_file, sheet, cols, rows, cols_index )
}
\author{
Pedro Guarderas, Andrés Lopez
}
\seealso{
\code{\link{Read.Utilities}}, \code{\link{Make.Decision.Tree}}
}

