% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.Utilities.R
\name{Read.Utilities}
\alias{Read.Utilities}
\title{Read utilities}
\usage{
Read.Utilities(file, script, lines, skip = 5)
}
\arguments{
\item{file}{standarized file with definitions}

\item{script}{output script where the utility functions are defined automatically}

\item{lines}{number lines to read in \code{file}}

\item{skip}{to read the \code{file} it had to \code{skip} a given number of lines}
}
\value{
Returns data table with definition of utility functions by range
}
\description{
Builds utility functions from definition standard
}
\details{
Details
}
\examples{
file<-'example/funciones_utilidad.txt'
sheet<-'example/funciones_utilidad.R'
nr<-15
skip<-5

}
\author{
Pedro Guarderas, Andrés Lopez
}
\seealso{
\code{\link{read_weights}}, \code{\link{eval_index}}
}

