% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stand.String.R
\name{Stand.String}
\alias{Stand.String}
\title{Standarize strings}
\usage{
Stand.String(x, chr = NULL, rep = NULL)
}
\arguments{
\item{x}{text to be standarized}

\item{chr}{character vector of replace characters}

\item{rep}{character vector of replacement characters}
}
\value{
Returns data table with definition of utility functions by range
}
\description{
Function to correct and standarize names, designed to eliminate special characters,
spaces and other prepositions.
}
\examples{
library( data.table )
library( stringr )
x<-c( 'H?Úàn with C@1_ad1', "M¡a/¬°r&ìa} *_the#-rot",
      'ju\%LIÖ a Pérs', '(S)tev\\nén\\t los cat\%$' )
y<-sapply( x, FUN = Stand.String )
names( y )<-NULL
}
\author{
Julio Andrade, Pedro Guarderas, Andrés Lopez
}

