% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corchar.R
\name{Stand.String}
\alias{Stand.String}
\title{Standarize strings}
\usage{
Stand.String(data, cols, chr = NULL, rep = NULL)
}
\arguments{
\item{data}{Data frame que contiene el texto a modificar.}

\item{cols}{Número de columna que contiene el texto a modificar.}

\item{chr}{Vector nulo de almacenamiento de caracteres por corregir.}

\item{rep}{Vector nulo de almacenamiento de caracteres de reemplazo.}
}
\value{
Returns data table with definition of utility functions by range
}
\description{
Función para corrección de caracteres, está diseñada para eliminar y/o reemplazar 
pronombres, preposiciones comunes, caracteres especiales, espacios.
}
\details{
Los pronombres, preposiciones comunes, caracteres especiales, espacios a 
reemplazar y/o eliminar son:  y/o ', ' las ', ' el ', ' para ', ' o ', ' en ', ' del ', ' de ', 
' la ', ' a ', ' por ', ' y ', '/', " ( )*", ' ', 'á', 'é', 'í', 'ó', 'ú', '-', '\\(', '\\)', 
'( )$', '_(_)*'.
}
\examples{
tst<-data.frame(a=c('HÚàn', 'Marìa','juLIO Pér', 'Stevén'), b=1:4)
cols<-1
correct_char2(tst, 1)
tst[,1]<-correct_char2(tst,1)

}
\author{
Julio Andrade, Pedro Guarderas, Andrés Lopez
}
\seealso{
\code{\link{owngsub}}, \code{\link{correct_char}}
}

