% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stand.String.R
\name{Stand.String}
\alias{Stand.String}
\title{Standarize strings}
\usage{
Stand.String(x, chr = NULL, rep = NULL)
}
\arguments{
\item{chr}{Vector nulo de almacenamiento de caracteres por corregir.}

\item{rep}{Vector nulo de almacenamiento de caracteres de reemplazo.}

\item{data}{Data frame que contiene el texto a modificar.}

\item{cols}{Número de columna que contiene el texto a modificar.}
}
\value{
Returns data table with definition of utility functions by range
}
\description{
Función para corrección de caracteres, está diseñada para eliminar y/o reemplazar 
pronombres, preposiciones comunes, caracteres especiales, espacios.
}
\details{
Los pronombres, preposiciones comunes, caracteres especiales y espacios.
}
\examples{
x<-c( 'H?Úàn with C@1_ad1', "M¡a/¬°r&ìa} *_the#-rot",
      'ju\%LIÖ a Pérs', '(S)tev\\nén\\t los cat\%$' )
y<-sapply( x, FUN = Stand.String )
names( y )<-NULL
}
\author{
Julio Andrade, Pedro Guarderas, Andrés Lopez
}
